workers/subtitle-generator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
boto3==1.29.7
python-multipart==0.0.6
openai-whisper==20231117
faster-whisper==1.0.0
torch>=2.1
torchaudio
numpy>=1.24.3
soundfile>=0.12.0
librosa>=0.10.0

